portfolio views is now a thing
