+++ TradePro_Reimagined/API/home/portfolio_historical_returns.html
@@ -1,6 +1,3 @@
 {% extends "home/portfolio.html" %}
 {% load static %}
 {% block body %}
-
-
-{% endblock %}
