+++ TradePro_Reimagined/API/templates/home/portfolio.html
@@ -44,7 +44,7 @@ which include seeing historical returns vs benchmark data -->
         Portfolio Options
     </button>
     <ul class="dropdown-menu" aria-labelledby="dropdownMenuButton1">
-        <li><a class="dropdown-item" href="{% url 'log_out' %}">Historical Returns</a></li>
+        <li><a class="dropdown-item" href="{% url 'portfolio/historical_returns' %}">Historical Returns</a></li>
         <!-- <li><a class="dropdown-item" href="{% url 'profile' %}">Profile</a></li> -->
 
         <!-- Trying to integrate javascript for MetaMask purposes -->
