+++ TradePro_Reimagined/API/urls.py
@@ -1,13 +1,14 @@
 from django.urls import path
 from . import views
 from .web3_views import web3_views
+from .portfolio_views import portfolio_views
 
 
 urlpatterns = [
 
         path("",views.Index.as_view(),name="index"),
         path("portfolio", views.Portfolio.as_view(), name="portfolio"),
-        path("portfolio/historical_returns", views.PortfolioHistoricalReturns.as_view(), name="portfolio/historical_returns"),
+        path("portfolio/historical_returns", portfolio_views.PortfolioHistoricalReturns.as_view(), name="portfolio/historical_returns"),
         path("log_in", views.log_in, name="log_in"),
         path("log_out", views.log_out, name="log_out"),
         path("open_an_account", views.open_an_account, name="open_an_account"),
